Building RowPlay uses RowSong
- Allows many Track extension properties to be make fileprivate.
- RowSong just stores the kind, and the kindSelect statement is now a computed property.
